Added DepositService and method createDeposit

+++ src/main/java/com/Projekt/Bankomat/Service/BankAccountService.java
@@ -2,12 +2,16 @@ package com.Projekt.Bankomat.Service;
 
 import com.Projekt.Bankomat.Enums.AccountType;
 import com.Projekt.Bankomat.Enums.CurrencyType;
+import com.Projekt.Bankomat.Enums.DepositType;
 import com.Projekt.Bankomat.Exceptions.BankAccountNotFoundException;
+import com.Projekt.Bankomat.Exceptions.InvalidAccountDeletionException;
+import com.Projekt.Bankomat.Exceptions.InvalidDepositException;
 import com.Projekt.Bankomat.Generators;
 import com.Projekt.Bankomat.Models.BankAccount;
 import com.Projekt.Bankomat.Repository.BankAccountRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigDecimal;
 import java.util.List;
@@ -34,10 +38,6 @@ public class BankAccountService implements IBankAccountService{
         return acc.get(0).getAccountNr();
     }
 
-    public List<BankAccount> getUserBankAccounts(String email){
-        return bankAccountRepo.getAllUserAccounts(email);
-    }
-    
     public boolean isPaymentValid(String zNrKonta, String doNrKonta, BigDecimal kwota, CurrencyType waluta){
         var zKonta = bankAccountRepo.findByAccountNr(zNrKonta)
                 .orElseThrow(() -> new BankAccountNotFoundException(zNrKonta));
@@ -55,15 +55,19 @@ public class BankAccountService implements IBankAccountService{
         bankAccountRepo.save(doKonta);
         return true;
     }
-
+    @Override
+    public List<BankAccount> getUserBankAccounts(String email){
+        return bankAccountRepo.getAllUserAccounts(email);
+    }
+    @Override
     public void deleteAccount(String nrKonta){
         var konto = bankAccountRepo.findByAccountNr(nrKonta).orElseThrow(() -> new BankAccountNotFoundException(nrKonta));
-        if(konto.getBalance().compareTo(BigDecimal.ZERO) > 0){
-            System.out.println("NA KONCIE POZOSTALY PIENIADZE, PRZELEJ JE PRZED USUNIECIEM KONTA");
-            throw new RuntimeException();
+        if(konto.getBalance().compareTo(BigDecimal.ZERO) > 0 || (!konto.getDeposits().isEmpty())){
+            throw new InvalidAccountDeletionException();
         }
         bankAccountRepo.delete(konto);
     }
+    @Override
     public void createAccount(String email, AccountType accountType, CurrencyType waluta){
         var konto = BankAccount.builder()
                 .accountId(UUID.randomUUID().toString())
